chore(suite): rename index files, export components directly
