+++ packages/suite/src/components/suite/index.tsx
@@ -43,9 +43,9 @@ import AccountFormCloseButton from './AccountFormCloseButton';
 import { TroubleshootingTips } from './TroubleshootingTips';
 import { ConnectDevicePrompt } from './ConnectDevicePrompt';
 import { DeviceButton } from './DeviceButton';
-import Coin from './Coin';
-import CoinsGroup from './CoinsGroup';
-import CoinsList from './CoinsGroup/CoinsList';
+import { Coin } from './Coin';
+import { CoinsGroup } from './CoinsGroup/CoinsGroup';
+import { CoinsList } from './CoinsGroup/CoinsList';
 import CharacterCount from './CharacterCount';
 import CloseButton from './CloseButton';
 import TooltipSymbol from './TooltipSymbol';
